adds readability to the code
